修改的代码

+++ cloud-provider-payment-8001/src/main/java/com/atguigu/springcloud/entities/Payment.java
@@ -10,7 +10,8 @@ import java.io.Serializable;
 @NoArgsConstructor
 @Data
 public class Payment implements Serializable {
-    private Integer id;
+
+    private Long id;
     private String serial;
 }
